Remove rust-toolchain.toml and change settings for measuring coverage
