+++ .github/workflows/release.yml
@@ -27,4 +27,5 @@ jobs:
       - name: Publish packages to crates.io
         uses: katyo/publish-crates@v2
         with:
+          path: ./crates/core
           registry-token: ${{ secrets.CARGO_REGISTRY_TOKEN }}
