+++ Cargo.toml
@@ -1,42 +1,21 @@
-[package]
-name         = "auto-palette"
-version      = "0.1.1"
-edition      = "2021"
-authors      = ["Tatsuya Maki <t28oogle@gmail.com>"]
-license      = "MIT"
-readme       = "README.md"
-repository   = "https://github.com/t28hub/auto-palette"
-description  = "🎨 A Rust library that extracts prominent color palettes from images automatically."
-rust-version = "1.75.0"
-categories   = ["multimedia::images", "graphics", "algorithms"]
-keywords     = ["palette", "color", "image", "color-palette", "color-extraction"]
+[workspace]
+resolver = "2"
+members  = ["crates/core"]
 
-[dependencies]
-image      = "0.25.1"
-num-traits = "0.2.18"
-rand       = "0.8.5"
-rand_distr = "0.4.3"
+[workspace.package]
+edition    = "2021"
+version    = "0.1.1"
+authors    = ["Tatsuya Maki <t28oogle@gmail.com>"]
+license    = "MIT"
+homepage   = "https://github.com/t28hub/auto-palette"
+repository = "https://github.com/t28hub/auto-palette"
 
-[dev-dependencies]
-reqwest = { version = "0.12.4", features = ["blocking"] }
-rstest  = "0.19.0"
+[profile.dev]
+opt-level = 3
 
-[[example]]
-name = "basic"
-path = "examples/basic.rs"
+[profile.test]
+opt-level = 3
 
-[[example]]
-name = "algorithm"
-path = "examples/algorithm.rs"
-
-[[example]]
-name = "theme"
-path = "examples/theme.rs"
-
-[[example]]
-name = "image_url"
-path = "examples/image_url.rs"
-
-[[example]]
-name = "image_path"
-path = "examples/image_path.rs"
+[profile.release]
+lto       = true
+opt-level = 's'
