+++ crates/core/examples/image_path.rs
@@ -1,29 +1,28 @@
 #![deny(warnings)]
 
-use std::{str::FromStr, time::Instant};
+use std::time::Instant;
 
-use auto_palette::{ImageData, Palette, Theme};
+use auto_palette::{ImageData, Palette};
 
-/// Extracts a palette from an image file and finds the dominant colors using the specified theme.
+/// Extracts a palette from an image file.
 ///
-/// The theme can be provided as a command line argument.
+/// The image path can be provided as a command line argument.
 /// ```sh
-/// cargo run --example theme --release -- vivid
+/// cargo run --example image_path --release -- tests/assets/holly-booth-hLZWGXy5akM-unsplash.jpg
 /// ```
 fn main() {
-    // Read the theme from the command line arguments
-    let theme = match std::env::args().nth(1) {
-        Some(name) => Theme::from_str(&name)
-            .map_err(|_| format!("Failed to parse the them '{}'", name))
-            .unwrap(),
-        None => {
-            println!("No theme provided, using the default theme");
-            Theme::Basic
-        }
-    };
+    // Read the image path from the command line arguments
+    let path = std::env::args().nth(1).unwrap_or_else(|| {
+        println!("No image path provided, using the default image path");
+        "./crates/core/tests/assets/holly-booth-hLZWGXy5akM-unsplash.jpg".into()
+    });
 
-    // Load the image data from the file
-    let image_data = ImageData::load("tests/assets/holly-booth-hLZWGXy5akM-unsplash.jpg").unwrap();
+    let image_data = ImageData::load(path.clone()).unwrap();
+    println!(
+        "Loaded the image with dimensions {}x{}",
+        image_data.width(),
+        image_data.height()
+    );
 
     // Extract the palette from the image data
     let start = Instant::now();
@@ -37,7 +36,7 @@ fn main() {
     );
 
     // Find the top 5 swatches in the palette
-    let swatches = palette.find_swatches_with_theme(5, theme);
+    let swatches = palette.find_swatches(5);
     println!("#  Color\tPosition\tPopulation");
     for (i, swatch) in swatches.iter().enumerate() {
         print!("{}  ", i + 1);
