+++ crates/core/src/algorithm.rs
