+++ crates/core/src/image/errors.rs
