+++ crates/core/src/image/mod.rs
