+++ crates/core/src/math/clustering/kmeans.rs
