+++ crates/core/src/math/mod.rs
