+++ crates/core/src/math/neighbors/linear.rs
