+++ crates/core/src/math/number.rs
