+++ crates/core/src/math/sampling.rs
