+++ crates/core/tests/assets/colors/black.png
