+++ crates/core/tests/assets/colors/blue.png
