+++ crates/core/tests/assets/colors/gray.png
