+++ crates/core/tests/assets/colors/green.png
