+++ crates/core/tests/assets/colors/red.png
