+++ crates/core/tests/assets/colors/transparent.png
