+++ crates/core/tests/assets/colors/white.png
