+++ crates/core/tests/assets/flags/kn.png
