+++ crates/core/tests/assets/flags/my.png
