+++ crates/core/tests/assets/flags/np.png
