+++ crates/core/tests/assets/olympic_rings.png
