+++ crates/core/tests/palette.rs
